// src/Client/Models/ManagerDetails.ts
export default class ManagerDetails {
  id: number = -1;
  first_name: string = "";
  last_name: string = "";
  email: string = "";
  client_img: string = "";
  active: boolean = true;
  timeline: boolean = true;
  campaign: boolean = true;
  rewards: boolean = true;
  leaderboard: boolean = true;
  calendar: boolean = true;
  role: number = -1;
}  